Signature: Make key management Actor specific
