+++ includes/model/class-application.php
@@ -7,8 +7,6 @@
 
 namespace Activitypub\Model;
 
-use WP_Query;
-use Activitypub\Signature;
 use Activitypub\Activity\Actor;
 use Activitypub\Collection\Actors;
 
@@ -166,7 +164,7 @@ class Application extends Actor {
 	 * @return string The published date.
 	 */
 	public function get_published() {
-		$first_post = new WP_Query(
+		$first_post = new \WP_Query(
 			array(
 				'orderby' => 'date',
 				'order'   => 'ASC',
@@ -219,7 +217,7 @@ class Application extends Actor {
 		return array(
 			'id'           => $this->get_id() . '#main-key',
 			'owner'        => $this->get_id(),
-			'publicKeyPem' => Signature::get_public_key_for( Actors::APPLICATION_USER_ID ),
+			'publicKeyPem' => Actors::get_public_key( Actors::APPLICATION_USER_ID ),
 		);
 	}
